TimeSync: Workaround empty responses (Pupil Capture)
We experienced during the experiment that sometimes
empty responses would be sent, which resulted in
the double.Parse method failing. To avoid the game
crashing in this case, we add a try catch and a if statement. the problem need to be more investigate this is for the moment a workaround.